requirements.txt
kafka-python
confluent-kafka
pandas
pyspark
fastavro
